# api_gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/donate": {
            "post": {
                "description": "Create a new donation registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registry"
                ],
                "summary": "Create a donation registry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Donation request data",
                        "name": "donation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRegistryReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pbRegistryRest.DonateResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/donated": {
            "get": {
                "description": "Get all registries for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registry"
                ],
                "summary": "Get all registries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by registry status",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pbRegistryRest.RegistriesResp"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/donated/{id}": {
            "get": {
                "description": "Get details of a specific registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registry"
                ],
                "summary": "Get registry detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbRegistryRest.DetailRegistryResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a registry by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registry"
                ],
                "summary": "Delete a registry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pbRegistryRest.DeleteResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateRegistryReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "donation_id": {
                    "type": "string"
                },
                "filter": {
                    "type": "string"
                }
            }
        },
...
        "pbRegistryRest.DeleteResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "pbRegistryRest.DetailRegistryResp": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "amount_to_pay": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "donation_id": {
                    "type": "string"
                },
                "donation_name": {
                    "type": "string"
                },
                "donor_id": {
                    "type": "integer"
                },
                "recipient_id": {
                    "type": "integer"
                },
                "recipient_name": {
                    "type": "string"
                },
                "registry_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "pbRegistryRest.DonateResp": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "donation_id": {
                    "type": "string"
                },
                "donor_id": {
                    "type": "integer"
                },
                "registry_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "pbRegistryRest.RegistriesResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pbRegistryRest.RegistryResp"
                    }
                }
            }
        },
        "pbRegistryRest.RegistryResp": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "donation_id": {
                    "type": "string"
                },
                "donor_id": {
                    "type": "integer"
                },
                "registry_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}